Add files via upload
This is a python program that uses the computer's built-in operating system to create a graphic user interface (GUI) that asks the user to pick a color and the program will generate the hexadecimal code corresponding to the selected color.

Please note: This program was used in conjunction with breezypythongui. In order for colorpicker to function, both of these files must be within the same working directory.